requirements.txt
Flask==3.0.3
SQLAlchemy==2.0.31
python-dotenv==1.0.1
google-genai==0.5.0
python-dotenv
